Move tree-planting into it's own class
The tree-planting endpoint is (I believe) the oldest endpoint in this
app and as such, it was built in the original-flavor style (all the code
in one endpoint method in app.rb).

As I've been refactoring & DRYing up the endpoints that I added, it
seemed like a good time to move this code over to the newer style (using
a subclass of a socrata base class) to reduce overall mental overhead
needed to understand the codebase.

I have a branch that will update the readme to reflect this change, but
I want to convert the other old-school endpoint (tow-away zones) over to
the new style first.