Humanize the tree planting data
I thought the fancy_title for the tree planting data sounded a bit
robotic, so I took the liberty of transforming the tree species and
site_info data into something slightly more conversational.

For the species, it seemed like the format from the API was "{scientific
plant name} :: {common name}". I swapped it to be "{common name}
({scientific name})", which seems more how an English speaker would
write it.

When the API returns us records with only one species name,
we just downcase it and remove "(s)".  Why? Because in all the cases I've
seen, if there is just one name given, its "Tree(s)" - and the sentence
sounds better if it just says "tree". Why do I do all the downcasing and
gsubbing? Why not just return "tree" directly? Mostly because I'm not
100% sure there isn't another case where we are only given one name. It
that situation, downcasing it and striping any "(s)"'s from it seems
like a sane default.

For the site_info text, I basically just made some guesses for all of
the different site_info's I saw in the most recent 100 records.
¯\_(ツ)_/¯